strbuf: Avoid memcpy() inside realloc() when growing with used == 0
If a buffer isn't used but is pre-allocated (e.g. was previously
allocated, then reset), avoid the memcpy() within realloc() by
allocating a brand new buffer and freeing the old one instead of
reallocating the buffer.

This is common on FastCGI requests: the strbuf is used to build the
params record payload, and then reused to receive the standard output
records.