Force alignment of struct lwan_handler_info
For some reason I don't yet know how, if the "route" member is added to
this struct, then find_handler() will segfault as there seems to be
some kind of discrepancy between the struct size and things in the
section.  Aligning the struct on an 8-byte boundary works around the
issue, although I have no clue why that happens.