Improve performance of url_decode()
This function was previously taking ~9.7% of instructions when handling
a request.  With this change, it now takes ~1.27%, making it over 7x
more efficient in the normal case where the request URL doesn't need to
be decoded.

The previous version wasn't that efficient because it worked roughly
like a branchy bytewise memcpy(), whereas the new version avoids copies
unless necessary.  Strings are still decoded in-place, but I imagine
there's still some opportunities for optimization, especially if vector
instructions are used directly rather than using string-handling functions
from the C library.

I haven't fuzzed this but since this is part of what's fuzzed by our
fuzzing infrastructure, I'm not really worried.