Add source code for https://smolsite.zip
This is a website that hosts websites within the URL: zip, base-64
encode the zip, and Lwan will serve that for you.  It's pretty much
useless, but it was fun to implement.

There are other things that could be cleaned up, including the
hardcoded HTML/JS and ZIP files within the C source code.

ZIP reading code needs extensive fuzzing too.