Use a template for the smolsite sample & build ZIP during build
This should be slightly more efficient than calling snprintf() to build
the main HTML with the iframe.  It's also quite a bit easier to update,
as one just edits smolfile.html and rebuild.

The sample smolsite is also generated during build, so updating it
requires editing index.html and rebuilding.