Tweak SO_ATTACH_REUSEPORT_CBPF usage
I've been playing with this for months and never found the reason
(more context in the comments I'm adding here) why this doesn't really
work the way it's intended to.  Instead of throwing away this patch,
let's just commit it and make this something for future me to continue
investigating when I'm in a better headspace.