Add FORTH implementation for a Forth Salon-compatible sample
This FORTH implementation has been written from scratch to be
compatible with the Forth Salon (https://forthsalon.appspot.com)
website, in order to use it in conjunction with LED Matrix panels to
create interactive art projects.  It should be 100% compatible,
although it hasn't been fully tested with a substantial corpus from FS.
Minimal fuzzing has been performed and it's stable enough for a first
version.

No integration with the Lwan build system or with the rgb-led-matrix
(https://github.com/hzeller/rpi-rgb-led-matrix) project has been
written yet; those will come later.  The idea is to have something
similar to the Forth Salon website, but editing a haiku would make it
show up in a LED matrix panel.