Fix @/! when stack is empty
Both loading from the memory (@) and storing (!) require popping the
address from the stack.  If it's empty, POP_D() will yield NaN -- making
the conversion to int32_t yield an out-of-bounds value. Handle this
condition.

This also gets rid of the STORE() and LOAD() macro, that were only used
inside one function each.