Use tail-call instruction dispatch for the forth runtime
This yields a ~2x performance boost in the benchmark.  This hasn't been
fully tested yet, and there's a good chance it'll blow up in your face.
This also needs some cleaning up; currently, we're generating an IR,
which is the old instruction set, and then converting it to this new scheme.
It was just easier to perform some of the transforms.