Macros to lazily-initialize globals: either once, or thread-local
This makes it easier to declare a lazily-initialized value by
implementing only a function that initializes and returns a value; the
rest is hidden by the macro.

For instance, to declare a lazily-initialized thread-local, one would
write:

	LWAN_LAZY_THREAD_LOCAL(bool, is_something_supported)
        {
             /* Perform computation to determine if something
              * is supported */
             return true;
        }

This function will be called only once per calling thread.  Likewise
with the lazy global macro (LWAN_LAZY_GLOBAL), where values are
initialized using pthread_once.