Add support for __attribute__((access))
This is a GCC extension that lets one mark how a function parameter
is used, especially if it is a pointer to a buffer with another parameter
used with the length.  You can specify, for instance, that a parameter
is read-only, giving a bit more protection against sloppy usage of const
and extra checks that require the value to be initialized before calling
the function.

Details:
https://gcc.gnu.org/onlinedocs/gcc/Common-Function-Attributes.html#index-access-function-attribute