Disable -Wstringop-truncation warning
Although I'd like to keep this warning enabled, let's disable it for
now, as it seems that, whem building in release mode with LTO, the
pragmas to ignore it are not considered.  All the current uses where
there's a warning are fine (the code knows that a NUL character may not
exist).